requirements.txt
APScheduler==3.7.0
beautifulsoup4==4.9.3
boto3==1.17.27
botocore==1.20.112
certifi==2022.9.24
chardet==4.0.0
click==8.1.3
dnspython==2.1.0
Flask==2.2.2
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.1
pause==0.3
pymongo==3.11.3
python-dateutil==2.8.2
python-dotenv==0.15.0
pytz==2022.5
requests==2.25.1
s3transfer==0.3.7
six==1.16.0
soupsieve==2.3.2.post1
tzlocal==2.1
urllib3==1.26.12
Werkzeug==2.2.2
